requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudscraper==1.2.56
colorama==0.4.4
Flask==1.1.2
flask-ipban==1.1.2
gunicorn==20.0.4
h11==0.12.0
html-text==0.5.2
html5lib==1.1
httpcore==0.12.3
httpx==0.16.1
hurry.filesize==0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
NHentai-API==0.0.15
opencv-python==4.5.1.48
pdf2image==1.14.0
Pillow==8.1.0
pyparsing==2.4.7
qrcode==6.1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
urllib3==1.26.3
Werkzeug==1.0.1
youtube-search-python==1.3.9
